feat/server-final

+++ backend/controller/goalController.js
@@ -1,13 +1,16 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 
-// @desc Get goals
+// @desc Get goals`
 // @route GET /api/goals
 // @access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-        const goals = await Goal.find();
+        const goals = await Goal.find({
+                user: req.user.id
+        });
 
         res.status(200).json(goals);
 });
@@ -22,7 +25,8 @@ const setGoal = asyncHandler(async (req, res) => {
                 throw new Error("Please add text field");
         }
         const goal = await Goal.create({
-                text: req.body.text
+                text: req.body.text,
+                user: req.user.id
         });
 
         res.status(200).json(goal);
@@ -38,7 +42,17 @@ const updateGoal = asyncHandler(async (req, res) => {
                 res.status(400);
                 throw new Error("Goal not found");
         }
+        const user = await User.findById(req.user.id);
 
+        if (!user) {
+                res.status(401);
+                throw new Error("User not found");
+        }
+
+        if (goal.user.toString() !== user.id) {
+                res.status(401);
+                throw new Error('User not authorized');
+        }
         const updatedGoal = await Goal.findByIdAndUpdate(
                 req.params.id,
                 req.body, { new: true }
@@ -58,6 +72,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
                 throw new Error("Goal not found");
         }
 
+        const user = await User.findById(req.user.id);
+
+        if (!user) {
+                res.status(401);
+                throw new Error("User not found");
+        }
+
+        if (goal.user.toString() !== user.id) {
+                res.status(401);
+                throw new Error('User not authorized');
+        }
         const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
 
         res.status(200).json(deletedGoal);
